Move parquet-related enums to Parquet namespace.
